programdetail dal sort fix
